requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
numpy==1.21.5
opencv-python==4.10.0.84
pandas==1.4.3
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
simplejson==3.19.2
six==1.16.0
urllib3==2.2.2
